Рефакторинг и мелкие исправления в протколе модбаса

+++ network/protocols/modbusprotocol.cpp
@@ -59,9 +59,9 @@ ModbusProtocol::DataVector ModbusProtocol::execute(const QByteArray& rxPackage, 
                 bytesCount = rxPackage.at(2);
                 if(rxPackage.size() == 5+bytesCount) {
                     for(quint16 i = 0; i < bytesCount/2; ++i) {
-                        quint16 lastWriteReg = (lastTxPackage.at(2) << 8) + lastTxPackage.at(3);
+                        quint16 lastWriteReg = (static_cast<quint8>(lastTxPackage.at(2)) << 8) + lastTxPackage.at(3);
                         reg = lastWriteReg+i;
-                        value = static_cast<quint16>((static_cast<quint16>(rxPackage.at(3+2*i)) << 8) + static_cast<quint16>(rxPackage.at(4+2*i)));
+                        value = static_cast<quint16>((static_cast<quint8>(rxPackage.at(3+2*i)) << 8) + static_cast<quint8>(rxPackage.at(4+2*i)));
                         resultVector.append({addr, reg, value});
                     }
                 } else {
